Limit entry note textarea to four lines when editing
